+++ eotsmanager/store/storedsigningrecord.go
@@ -1,17 +1,33 @@
 package store
 
-import "github.com/babylonlabs-io/finality-provider/eotsmanager/proto"
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	"github.com/babylonlabs-io/finality-provider/eotsmanager/proto"
+)
 
 type SigningRecord struct {
-	BlockHash []byte // The hash of the block.
-	PublicKey []byte // The public key used for signing.
-	Signature []byte // The signature of the block.
-	Timestamp int64  // The timestamp of the signing operation, in Unix seconds.
+	Msg       []byte // The message that the signature is signed over.
+	Signature []byte
+	Timestamp int64 // The timestamp of the signing operation, in Unix seconds.
 }
 
 func (s *SigningRecord) FromProto(sr *proto.SigningRecord) {
-	s.PublicKey = sr.PublicKey
-	s.BlockHash = sr.BlockHash
+	s.Msg = sr.Msg
 	s.Timestamp = sr.Timestamp
-	s.Signature = sr.Signature
+	s.Signature = sr.EotsSig
+}
+
+// the record key is (chainID || pk || height)
+func getSignRecordKey(chainID, pk []byte, height uint64) []byte {
+	// Convert height to bytes
+	heightBytes := sdk.Uint64ToBigEndian(height)
+
+	// Concatenate all components to create the key
+	key := make([]byte, 0, len(pk)+len(chainID)+len(heightBytes))
+	key = append(key, chainID...)
+	key = append(key, pk...)
+	key = append(key, heightBytes...)
+
+	return key
 }